+++ src/index.js
@@ -2,69 +2,10 @@ import './style.css';
 import './assets/img/square.svg';
 import './assets/img/more-vertical.svg';
 import './assets/img/bin.svg';
+import Lists from './module/listFuns.js';
 
 // add
 const todoEl = document.querySelector('.todo-list');
-
-class Lists {
-  constructor() {
-    this.tdlists = JSON.parse(localStorage.getItem('tdlists')) || [];
-  }
-
-  delList(item, idx) {
-    this.tdlists = JSON.parse(localStorage.getItem('tdlists')) || [];
-    const tdIdx = this.tdlists[idx].index;
-    this.tdlists.splice(tdIdx, 1);
-    this.tdlists.forEach((tdlist, tdnum) => {
-      tdlist.index = tdnum;
-    });
-
-    localStorage.setItem('tdlists', JSON.stringify(this.tdlists));
-    this.showList();
-  }
-
-  showList() {
-    todoEl.innerHTML = '';
-    this.tdlists.forEach((tdlist) => {
-      const lists = document.createElement('div');
-      lists.classList = 'lists';
-      const checkBox = document.createElement('img');
-      checkBox.src = './assets/img/square.svg';
-      checkBox.classList = 'check-box';
-      checkBox.alt = 'square';
-
-      const todoText = document.createElement('input');
-      todoText.type = 'text';
-      todoText.name = 'text';
-      todoText.classList = 'todo-text';
-      todoText.value = tdlist.description;
-      todoText.disabled = true;
-
-      const more = document.createElement('img');
-      more.src = './assets/img/more-vertical.svg';
-      more.classList = 'more';
-      more.alt = 'more-vertical';
-
-      lists.append(checkBox, todoText, more);
-      todoEl.appendChild(lists);
-    });
-  }
-
-  storeLocal(td) {
-    this.tdlists.push({
-      description: td,
-      completed: false,
-      index: this.tdlists.length,
-    });
-    localStorage.setItem('tdlists', JSON.stringify(this.tdlists));
-  }
-
-  edit(val, idx) {
-    this.tdlists = JSON.parse(localStorage.getItem('tdlists')) || [];
-    this.tdlists[idx].description = val;
-    localStorage.setItem('tdlists', JSON.stringify(this.tdlists));
-  }
-}
 
 const listFuns = new Lists();
 listFuns.showList();
@@ -119,6 +60,8 @@ const editTodo = () => {
       if (e.key === 'Enter' || e.key === 'Tab') textnormal(text);
     });
   });
+
+  localStorage.setItem('tdlists', JSON.stringify(listFuns.tdlists));
 };
 
 // submit
